Put all of the choropleth related layout components into one file
