--- src/components-styled/choropleth/choropleth-chart.tsx
@@ -1,9 +1,0 @@
-import css from '@styled-system/css';
-import styled from 'styled-components';
-import { Box } from '../base';
-
-export const ChoroplethChart = styled(Box)(
-  css({
-    gridArea: 'c',
-  })
-);
