--- src/components-styled/choropleth/choropleth-header.tsx
@@ -1,10 +1,0 @@
-import css from '@styled-system/css';
-import styled from 'styled-components';
-import { Box } from '../base';
-
-export const ChoroplethHeader = styled(Box)(
-  css({
-    gridArea: 'a',
-    mb: [0, 2],
-  })
-);
