+++ src/components-styled/layout/choropleth.tsx
@@ -2,6 +2,28 @@ import css from '@styled-system/css';
 import styled from 'styled-components';
 import { Box } from '../base';
 
+export const ChoroplethChart = styled(Box)(
+  css({
+    gridArea: 'c',
+  })
+);
+
+export const ChoroplethHeader = styled(Box)(
+  css({
+    gridArea: 'a',
+    mb: [0, 2],
+  })
+);
+
+export const ChoroplethLegend = styled(Box)(
+  css({
+    gridArea: 'b',
+    display: 'flex',
+    flexDirection: 'column',
+    alignItems: ['center', 'center', 'center', 'flex-start'],
+  })
+);
+
 export const ChoroplethSection = styled(Box)(
   css({
     bg: 'white',
